requirements.txt
streamlit
praw
pandas
matplotlib
seaborn
vaderSentiment
wordcloud
torch
transformers
python-dotenv
